+++ src/main/java/com/codesquad/blackjack/domain/Chip.java
@@ -1,12 +1,21 @@
 package com.codesquad.blackjack.domain;
 
+import com.codesquad.blackjack.domain.player.Gamer;
+
+import javax.persistence.*;
 import java.util.Objects;
 
+@Entity
 public class Chip {
     public static final int ZERO = 0;
     private static final int TWO = 2;
     private static final double ONE_POINT_FIVE = 1.5;
 
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private long id;
+
+    @Column
     private final int amount;
 
     public Chip(int amount) {
@@ -55,5 +64,13 @@ public class Chip {
         return "Chip{" +
                 "amount=" + amount +
                 '}';
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public int getAmount() {
+        return amount;
     }
 }
