+++ src/main/java/com/codesquad/blackjack/webPackage/domain/User.java
@@ -1,70 +1,70 @@
-package com.codesquad.blackjack.webPackage.domain;
-
-
-import javax.persistence.*;
-import javax.validation.constraints.Size;
-
-@Entity
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
-
-    @Size(min = 3, max = 20)
-    @Column(unique = true, nullable = false, length = 20)
-    private String userId;
-
-    @Size(min = 6, max = 20)
-    @Column(nullable = false, length = 20)
-    private String password;
-
-    @Size(min = 3, max = 20)
-    @Column(nullable = false, length = 20)
-    private String name;
-
-    public User() {
-    }
-
-    public User(long id, String userId, String password, String name) {
-        this.id = id;
-        this.userId = userId;
-        this.password = password;
-        this.name = name;
-    }
-
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public boolean matchPassword(String password) {
-        return this.password.equals(password);
-    }
-}
+//package com.codesquad.blackjack.webPackage.domain;
+//
+//
+//import javax.persistence.*;
+//import javax.validation.constraints.Size;
+//
+//@Entity
+//public class User {
+//    @Id
+//    @GeneratedValue(strategy = GenerationType.IDENTITY)
+//    private long id;
+//
+//    @Size(min = 3, max = 20)
+//    @Column(unique = true, nullable = false, length = 20)
+//    private String userId;
+//
+//    @Size(min = 6, max = 20)
+//    @Column(nullable = false, length = 20)
+//    private String password;
+//
+//    @Size(min = 3, max = 20)
+//    @Column(nullable = false, length = 20)
+//    private String name;
+//
+//    public User() {
+//    }
+//
+//    public User(long id, String userId, String password, String name) {
+//        this.id = id;
+//        this.userId = userId;
+//        this.password = password;
+//        this.name = name;
+//    }
+//
+//    public long getId() {
+//        return id;
+//    }
+//
+//    public void setId(long id) {
+//        this.id = id;
+//    }
+//
+//    public String getUserId() {
+//        return userId;
+//    }
+//
+//    public void setUserId(String userId) {
+//        this.userId = userId;
+//    }
+//
+//    public String getPassword() {
+//        return password;
+//    }
+//
+//    public void setPassword(String password) {
+//        this.password = password;
+//    }
+//
+//    public String getName() {
+//        return name;
+//    }
+//
+//    public void setName(String name) {
+//        this.name = name;
+//    }
+//
+//    public boolean matchPassword(String password) {
+//        return this.password.equals(password);
+//    }
+//}
