+++ src/main/java/com/codesquad/blackjack/webPackage/domain/UserRepository.java
@@ -1,9 +1,9 @@
-package com.codesquad.blackjack.webPackage.domain;
-
-import org.springframework.data.repository.CrudRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends CrudRepository<User, Long> {
-    Optional<User> findByUserId(String userId);
-}
+//package com.codesquad.blackjack.webPackage.domain;
+//
+//import org.springframework.data.repository.CrudRepository;
+//
+//import java.util.Optional;
+//
+//public interface UserRepository extends CrudRepository<User, Long> {
+//    Optional<User> findByUserId(String userId);
+//}
