+++ src/main/resources/import.sql
@@ -1,2 +1,6 @@
-INSERT INTO user (id, user_id, password, name) values (1, 'finn', '1234', '동선');
-INSERT INTO user (id, user_id, password, name) values (2, 'zingo', '1234', '진산');
+
+INSERT INTO Gamer (id, user_id, password, name, chip_id) values (1, 'finn', '123456', '동선', 1);
+INSERT INTO Gamer (id, user_id, password, name, chip_id) values (2, 'zingo', '123456', '진산', 2);
+
+INSERT INTO Chip (id, amount) values (1, 500);
+INSERT INTO Chip (id, amount) values (2, 1000000);
